Add gitlab fully quialified domain name.
By defining the variable `gitlab_fqdn` it is easier to redefine
the variables:

  - 'gitlab_external_url
  - 'gitlab_ssl_certificate'
  - 'gitlab_ssl_certificate_key'

Not only that. When using the Ansible role `gitlab` in conjunction
with `gitlab-runner`, this method allows to define:

  - 'gitlab_runner_coordinator_url'
  - 'gitlab_runner_tls_ca_file'

This way by over-writing one variable, it is possible to redefine
five variables in two roles, simplifying the `groups_vars/all` file
and the risk of making mistakes.